+++ src/Pim/Attribute/AttributeService.php
@@ -7,6 +7,7 @@ declare(strict_types=1);
 
 namespace Wizaplace\SDK\Pim\Attribute;
 
+use GuzzleHttp\RequestOptions;
 use Wizaplace\SDK\AbstractService;
 
 final class AttributeService extends AbstractService
@@ -51,5 +52,27 @@ final class AttributeService extends AbstractService
         return array_map(static function (array $attributeData): Attribute {
             return new Attribute($attributeData);
         }, $data);
+    }
+
+    public function setProductAttributeValue(int $productId, int $attributeId, string $value): void
+    {
+        $this->client->mustBeAuthenticated();
+
+        $this->client->put("products/${productId}/features/${attributeId}", [
+            RequestOptions::JSON => [
+                'value_str' => $value,
+            ],
+        ]);
+    }
+
+    public function setProductAttributeVariants(int $productId, int $attributeId, array $variantIds): void
+    {
+        $this->client->mustBeAuthenticated();
+
+        $this->client->put("products/${productId}/features/${attributeId}", [
+            RequestOptions::JSON => [
+                'variant_id' => $variantIds,
+            ],
+        ]);
     }
 }