+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 - Added `\Wizaplace\SDK\Vendor\Order\OrderService::listOrders` and associated classes
 - Added `\Wizaplace\SDK\Pim\Attribute\AttributeService::getProductAttributes` and associated classes
 - Added `\Wizaplace\SDK\Pim\Attribute\AttributeService::getProductAttribute` and associated classes
+- Added `\Wizaplace\SDK\Pim\Attribute\AttributeService::setProductAttributeValue`
+- Added `\Wizaplace\SDK\Pim\Attribute\AttributeService::setProductAttributeVariants`
 - Added `\Wizaplace\SDK\Pim\Attribute\AttributeService::getCategoryAttributes` and associated classes
 - Added `\Wizaplace\SDK\Pim\Option\OptionService::getCategoryOptions` and associated classes